refactor features into json db
